refatcor: rename components
