+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
-import Layout from "../components/layout/layout"
 import {StaticImage} from "gatsby-plugin-image"
 import { Accordion} from "react-bootstrap"
+import Layout from "../components/layout"
 
 // markup
 const IndexPage = () => {
