+++ src/pages/{mdx.slug}.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { graphql } from 'gatsby';
 import { MDXRenderer } from 'gatsby-plugin-mdx';
-import Layout from '../components/layout/layout';
+import Layout from '../components/layout';
 
 const Page = ({ data }) => {
     const post = data.mdx;
